pyproject.toml
[tool.poetry]
name = "repl-nix-career-app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "repl_nix_career_app"}]

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.3"
snowflake-connector-python = "^3.12.0"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
